requirements.txt
Django>=5.1,<6
djangorestframework>=3.15,<4
strawberry-graphql>=0.289.0,<1
strawberry-graphql-django>=0.73.1,<1
channels>=4.1,<5
daphne>=4.1,<5
python-dotenv>=1.0,<2
